bug:
If a force_install_date in InstallInfo was in the past, logouthelper will
advance it to NOW + 60 minutes and start ignoring InstallInfo upcoming
force_install_date values.

If the admin revises the old force_install_date to the future after
logouthelper has already run, logouthelper does not load the new
force_install_date and sticks with its new value.

This patch fixes this.