Automatic Apple Updates
This adds support for automatic unattended installation of Apple updates.  It is intended to cut down on the need for creating pkginfos just to add unattended_install keys.

Setting `AutomaticAppleUpdates` to true will cause all Apple updates which do not require a logout or restart to be treated as unattended, getting the blocking applications list from the Apple distribution.

It is currently limited to just Yosemite and newer, though it can likely be extended down to Mavericks.  The behavior is similar to when "Install OS X updates" is set to true in the App Store's preferences.